pyproject.toml
[project]
authors = [{name = "Junyuan DENG"},{name="Heng LI"}]
dependencies = [
    "numpy<2",
    "Pillow",
    "huggingface_hub",
    "einops",
    "safetensors",
    "opencv-python",
    "torch>=2.3.1",
    "torchvision>=0.18.1",
    "numpy==1.26.1",
    "evo",
    "plyfile",
    #"python-opencv",
]
name = "sailrecon"
requires-python = ">= 3.10"
version = "0.0.1"

[project.optional-dependencies]
demo = [
    "gradio>=5.17.1",
    "viser>=0.2.23",
    "tqdm",
    "hydra-core",
    "omegaconf",
    "opencv-python",
    "scipy",
    "onnxruntime",
    "requests",
    "trimesh",
    "matplotlib",
]

# Using setuptools as the build backend
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

# setuptools configuration
[tool.setuptools.packages.find]
where = ["."]
include = ["sailrecon*"]

# Pixi configuration
[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
sailrecon = { path = ".", editable = true }

[tool.pixi.environments]
default = { solve-group = "default" }
demo = { features = ["demo"], solve-group = "default" }

[tool.pixi.tasks]
